# openapi.yaml
openapi: 3.0.1
info:
  title: GOMP
  version: "1.0"
servers:
  - url: /api/v1
paths:
...
  /auth:
    post:
      tags: [app]
      description: perform user authentication
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/credentials"
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authenticationResponse"
        401:
          description: Unauthorized
      x-codegen-request-body-name: credentials
  /auth/refresh:
    get:
      tags: [app]
      description: refresh the auth token given the existing token
      operationId: refreshToken
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/authenticationResponse"
        401:
          description: Unauthorized
      security:
        - Bearer: []
...
components:
  schemas:
    authenticationResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: "./models.yaml#/components/schemas/user"
    credentials:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT